# tests/data/openapi.json
{
  "components": {
    "schemas": {
...
      "FindingOut": {
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "duplicate_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duplicate Of"
          },
          "file_path": {
            "default": "",
            "title": "File Path",
            "type": "string"
          },
          "finding_id": {
            "title": "Finding Id",
            "type": "string"
          },
          "line_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line End"
          },
          "line_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line Start"
          },
          "poc_test": {
            "default": "",
            "title": "Poc Test",
            "type": "string"
          },
          "proposed_patch": {
            "default": "",
            "title": "Proposed Patch",
            "type": "string"
          },
          "provenance": {
            "additionalProperties": true,
            "title": "Provenance",
            "type": "object"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "severity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Severity"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          },
          "stable_key": {
            "title": "Stable Key",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "threat_model": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ThreatModel"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threat Model"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          },
          "validation_trail": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Trail"
          }
        },
        "required": [
          "finding_id",
          "run_id",
          "stable_key",
          "status"
        ],
        "title": "FindingOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_FindingOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FindingOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[FindingOut]",
        "type": "object"
      },
...
      "Page_WishOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WishOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[WishOut]",
        "type": "object"
      },
...
      "Severity": {
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "Severity",
        "type": "string"
      },
...
      "ThreatModel": {
        "properties": {
          "assumption_broken": {
            "description": "e.g. 'length field validated before use'",
            "title": "Assumption Broken",
            "type": "string"
          },
          "attacker": {
            "description": "e.g. 'unauthenticated remote client'",
            "title": "Attacker",
            "type": "string"
          },
          "boundary_crossed": {
            "description": "e.g. 'network input -> parser memory'",
            "title": "Boundary Crossed",
            "type": "string"
          }
        },
        "required": [
          "attacker",
          "boundary_crossed",
          "assumption_broken"
        ],
        "title": "ThreatModel",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ValidationOut": {
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "model": {
            "default": "",
            "title": "Model",
            "type": "string"
          },
          "pass_name": {
            "title": "Pass Name",
            "type": "string"
          },
          "prompt_version": {
            "default": "",
            "title": "Prompt Version",
            "type": "string"
          },
          "reasoning": {
            "default": "",
            "title": "Reasoning",
            "type": "string"
          },
          "response_class": {
            "default": "",
            "title": "Response Class",
            "type": "string"
          },
          "verdict": {
            "title": "Verdict",
            "type": "string"
          }
        },
        "required": [
          "pass_name",
          "verdict"
        ],
        "title": "ValidationOut",
        "type": "object"
      },
      "WishOut": {
        "properties": {
          "blocked_task_id": {
            "title": "Blocked Task Id",
            "type": "string"
          },
          "context": {
            "default": "",
            "title": "Context",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "need": {
            "title": "Need",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "run_id",
          "blocked_task_id",
          "need",
          "status"
        ],
        "title": "WishOut",
        "type": "object"
      },
      "WishResolveIn": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "title": "WishResolveIn",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Read-first access to runs, findings, reports and run artifacts.",
    "title": "Crucible API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/findings": {
      "get": {
        "operationId": "find_by_stable_key_findings_get",
        "parameters": [
          {
            "in": "query",
            "name": "stable_key",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Stable Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_FindingOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Find By Stable Key",
        "tags": [
          "findings"
        ]
      }
    },
    "/findings/{finding_id}": {
      "get": {
        "operationId": "get_finding_findings__finding_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Finding",
        "tags": [
          "findings"
        ]
      }
    },
    "/findings/{finding_id}/validations": {
      "get": {
        "operationId": "get_validations_findings__finding_id__validations_get",
        "parameters": [
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ValidationOut"
                  },
                  "title": "Response Get Validations Findings  Finding Id  Validations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Validations",
        "tags": [
          "findings"
        ]
      }
    },
...
    "/runs/{run_id}/findings": {
      "get": {
        "operationId": "list_run_findings_runs__run_id__findings_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "in": "query",
            "name": "min_severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Severity"
            }
          },
          {
            "in": "query",
            "name": "attack_class",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Attack Class"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "severity",
              "pattern": "^(severity|created_at)$",
              "title": "Order",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_FindingOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Run Findings",
        "tags": [
          "findings"
        ]
      }
    },
    "/runs/{run_id}/findings/upheld": {
      "get": {
        "operationId": "list_upheld_runs__run_id__findings_upheld_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_FindingOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Upheld",
        "tags": [
          "findings"
        ]
      }
    },
...
    "/runs/{run_id}/wishes": {
      "get": {
        "operationId": "list_run_wishes_runs__run_id__wishes_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_WishOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Run Wishes",
        "tags": [
          "wishlist"
        ]
      }
    },
    "/wishes": {
      "get": {
        "operationId": "list_wishes_wishes_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "open",
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_WishOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Wishes",
        "tags": [
          "wishlist"
        ]
      }
    },
    "/wishes/{wish_id}": {
      "get": {
        "operationId": "get_wish_wishes__wish_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "wish_id",
            "required": true,
            "schema": {
              "title": "Wish Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WishOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Wish",
        "tags": [
          "wishlist"
        ]
      }
    },
    "/wishes/{wish_id}/resolve": {
      "post": {
        "description": "Mark a wish resolved. Does NOT itself re-queue the blocked task \u2014 that\nhappens on the next `crucible run --resume <run_id>` (specs \u00a79.3).",
        "operationId": "resolve_wish_wishes__wish_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "wish_id",
            "required": true,
            "schema": {
              "title": "Wish Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/WishResolveIn"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WishOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Wish",
        "tags": [
          "wishlist"
        ]
      }
    }
  }
}